+++ .gitignore
@@ -1,1 +1,3 @@
 mysite/instance
+mysite/mainapp_v0.py
+mysite/pymodule/__pycache__
