+++ mysite/templates/change_password.html
@@ -5,7 +5,7 @@
 {% block content %}
     <h2>更改密码</h2>
     
-    <form action="{{ url_for('change_password') }}" method="post">
+    <form action="{{ url_for('auth.change_password') }}" method="post">
         <div class="form-group">
             <label for="old_password">旧密码</label>
             <input type="password" id="old_password" name="old_password" required>
@@ -23,7 +23,7 @@
         
         <div class="form-actions">
             <button type="submit" class="btn btn-primary">更改密码</button>
-            <a href="{{ url_for('profile') }}" class="btn btn-secondary">返回个人信息</a>
+            <a href="{{ url_for('auth.profile') }}" class="btn btn-secondary">返回个人信息</a>
         </div>
     </form>
 {% endblock %}
