+++ mysite/templates/edit_order.html
@@ -6,7 +6,7 @@
 <div class="main-content">
     <h1>修改订单</h1>
 
-    <form action="{{ url_for('edit_order', order_id=order.id) }}" method="post">
+    <form action="{{ url_for('order.edit_order', order_id=order.id) }}" method="post">
         <div class="form-group">
             <label for="category">类目</label>
             <input type="text" class="form-control" id="category" name="category" value="{{ order.category }}" required>
@@ -32,7 +32,7 @@
             <input type="text" class="form-control" id="contact" name="contact" value="{{ order.contact }}" required>
         </div>
         <button type="submit" class="btn btn-primary">保存修改</button>
-        <a href="{{ url_for('view_order', order_id=order.id) }}" class="btn btn-secondary">取消</a>
+        <a href="{{ url_for('order.view_order', order_id=order.id) }}" class="btn btn-secondary">取消</a>
     </form>
 </div>
 {% endblock %}
