+++ mysite/templates/list_order.html
@@ -5,10 +5,10 @@
 {% block content %}
 <div class="main-content">
     <h1>订单列表</h1>
-    <a href="{{ url_for('submit_order') }}" class="btn btn-primary mb-3">发布新订单</a>
+    <a href="{{ url_for('order.submit_order') }}" class="btn btn-primary mb-3">发布新订单</a>
 
     {# Filter Form #}
-    <form action="{{ url_for('list_order') }}" method="get" class="form-inline mb-3">
+    <form action="{{ url_for('order.list_order') }}" method="get" class="form-inline mb-3">
         <div class="form-group mr-2">
             <label for="category_filter" class="mr-2">按类目筛选:</label>
             <select class="form-control" id="category_filter" name="category">
@@ -28,7 +28,7 @@
             </select>
         </div>
         <button type="submit" class="btn btn-secondary">筛选</button>
-        <a href="{{ url_for('list_order') }}" class="btn btn-outline-secondary ml-2">清除筛选</a>
+        <a href="{{ url_for('order.list_order') }}" class="btn btn-outline-secondary ml-2">清除筛选</a>
     </form>
 
 
@@ -43,11 +43,11 @@
                     <p class="card-text"><strong>工期:</strong> {{ order.period }}</p>
                     <p class="card-text"><strong>发布人:</strong> {{ order.submiter.username }}</p> {# Access submiter's username #}
                     <p class="card-text"><strong>发布时间:</strong> {{ order.when_submitted.strftime('%Y-%m-%d %H:%M') }}</p> {# Format datetime #}
-                    <a href="{{ url_for('view_order', order_id=order.id) }}" class="btn btn-info btn-sm">查看详情</a>
+                    <a href="{{ url_for('order.view_order', order_id=order.id) }}" class="btn btn-info btn-sm">查看详情</a>
                     {% if session.get('user_id') == order.submiter_id %} {# Check if current user is the submiter #}
-                        <a href="{{ url_for('edit_order', order_id=order.id) }}" class="btn btn-warning btn-sm">修改</a>
+                        <a href="{{ url_for('order.edit_order', order_id=order.id) }}" class="btn btn-warning btn-sm">修改</a>
                         {# Use a form for delete button for safety #}
-                        <form action="{{ url_for('delete_order', order_id=order.id) }}" method="post" style="display: inline;">
+                        <form action="{{ url_for('order.delete_order', order_id=order.id) }}" method="post" style="display: inline;">
                             <button type="submit" class="btn btn-danger btn-sm" onclick="return confirm('确定要删除此订单吗？');">删除</button>
                         </form>
                     {% endif %}
